correct db

+++ migrations/20180703110436_locations.js
@@ -2,8 +2,8 @@ exports.up = function (knex, Promise) {
     return knex.schema.createTable('locations', (table) => {
         table.increments()
         table.text('name')
-        table.decimal('lat')
-        table.decimal('lng')
+        table.float('lat')
+        table.float('lng')
         table.text('hazards')
         table.integer('holes')
     })
